requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.4.26
click==8.2.1
discord.py==2.5.2
flake8==7.2.0
frozenlist==1.7.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
multidict==6.4.4
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
PySocks==1.7.1
pytest==8.4.0
pytest-asyncio==1.0.0
python-dotenv==1.1.0
pytz==2025.2
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
